+++ CHANGES.md
@@ -8,6 +8,7 @@ Changelog of threedi-modelchecker
 - Renamed some methods of ThreediModelChecker.
 - Added basic to the 3di model schema: checks if the model has the latest migration 
   applied and raises an error if not.
+- Rewrote CrossSectionShape check to no longer use regex and added it to config.
 
 
 0.1 (2019-06-04)